+++ src/components/Career/CareersDetail/component.tsx
