+++ src/components/Career/CareersDetail/index.tsx
