+++ src/components/Career/CareersDetail/interface.tsx
